Refactoring: move stuff from main to lib
